Moving Level constructor definition in .cpp
